+++ innopoints/views/__init__.py
@@ -1,4 +1,5 @@
 from .authentication import *
 from .project import *
+from .activity import *
 
 from .views import *  # temporary, until all views have been moved
