+++ innopoints/views/views.py
@@ -15,20 +15,16 @@ from innopoints.extensions import db
 from innopoints.blueprints import api
 import innopoints.file_manager_s3 as file_manager
 from innopoints.models import (
-    Activity,
     Color,
     Competence,
     Product,
-    Project,
     Size,
     StaticFile,
     StockChange,
     StockChangeStatus,
     Variety,
-    IPTS_PER_HOUR,
 )
 from innopoints.schemas import (
-    ActivitySchema,
     ColorSchema,
     CompetenceSchema,
     ProductSchema,
@@ -39,107 +35,6 @@ from innopoints.schemas import (
 
 ALLOWED_MIMETYPES = {'image/jpeg', 'image/png', 'image/webp'}
 NO_PAYLOAD = ('', 204)
-
-
-# ----- Activity -----
-
-@api.route('/projects/<int:project_id>/activity', methods=['POST'])
-@login_required
-def create_activity(project_id):
-    """Create a new activity to an existing project."""
-    if not request.is_json:
-        abort(400, {'message': 'The request should be in JSON.'})
-
-    project = Project.query.get_or_404(project_id)
-    if not current_user.is_admin and current_user not in project.moderators:
-        abort(401)
-
-    in_schema = ActivitySchema(exclude=('id', 'project', 'applications', 'notifications'))
-
-    try:
-        new_activity = in_schema.load(request.json)
-    except ValidationError as err:
-        abort(400, {'message': err.messages})
-
-    new_activity.project = project
-
-    try:
-        db.session.add(new_activity)
-        db.session.commit()
-    except IntegrityError as err:
-        db.session.rollback()
-        print(err)  # TODO: replace with proper logging
-        abort(400, {'message': 'Data integrity violated.'})
-
-    out_schema = ActivitySchema(exclude=('notifications', 'existing_application'),
-                                context={'user': current_user})
-    return out_schema.jsonify(new_activity)
-
-
-class ActivityAPI(MethodView):
-    """REST views for a particular instance of an Activity model."""
-
-    @login_required
-    def patch(self, project_id, activity_id):
-        """Edit the activity."""
-        if not request.is_json:
-            abort(400, {'message': 'The request should be in JSON.'})
-
-        project = Project.query.get_or_404(project_id)
-        if not current_user.is_admin and current_user not in project.moderators:
-            abort(401)
-
-        activity = Activity.query.get_or_404(activity_id)
-        if activity.project != project:
-            abort(400, {'message': 'The specified project and activity are unrelated.'})
-
-        in_schema = ActivitySchema(exclude=('id', 'project', 'applications', 'notifications'))
-
-        try:
-            updated_activity = in_schema.load(request.json, instance=activity, partial=True)
-        except ValidationError as err:
-            abort(400, {'message': err.messages})
-
-        if not activity.fixed_reward and activity.reward_rate != IPTS_PER_HOUR:
-            abort(400, {'message': 'The reward rate for hourly activities may not be changed.'})
-
-        try:
-            db.session.add(updated_activity)
-            db.session.commit()
-        except IntegrityError as err:
-            db.session.rollback()
-            print(err)  # TODO: replace with proper logging
-            abort(400, {'message': 'Data integrity violated.'})
-
-        out_schema = ActivitySchema(exclude=('notifications', 'existing_application'),
-                                    context={'user': current_user})
-        return out_schema.jsonify(updated_activity)
-
-    @login_required
-    def delete(self, project_id, activity_id):
-        """Delete the activity."""
-        project = Project.query.get_or_404(project_id)
-        if not current_user.is_admin and current_user not in project.moderators:
-            abort(401)
-
-        activity = Activity.query.get_or_404(activity_id)
-        if activity.project != project:
-            abort(400, {'message': 'The specified project and activity are unrelated.'})
-
-        try:
-            db.session.delete(activity)
-            db.session.commit()
-        except IntegrityError as err:
-            db.session.rollback()
-            print(err)  # TODO: replace with proper logging
-            abort(400, {'message': 'Data integrity violated.'})
-        return NO_PAYLOAD
-
-
-activity_api = ActivityAPI.as_view('activity_api')
-api.add_url_rule('/projects/<int:project_id>/activity/<int:activity_id>',
-                 view_func=activity_api,
-                 methods=('PATCH', 'DELETE'))
 
 
 # ----- Product -----
